extract choropleth into selectors
